requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0
Pillow==7.1.2
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.7
djangorestframework==3.11.0
django-filter==2.2.0
djangorestframework-gis==0.15

# JWT
djangorestframework-simplejwt==4.4.0
pyjwt==1.7.1

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Celery
redis>=2.10.6, < 3
django-redis==4.10.0
celery==4.2.1
flower==0.9.2

# Channeles
channels==2.4.0
channels-redis==2.4.2

# Cors
django-cors-headers==3.2.1

# Facebook
facebook-sdk==3.1.0

# Notifications FCM
fcm-django==0.3.4

# Swagger
django-swagger-ui==0.1.11
django-rest-swagger==2.2.0
pyyaml==5.3.1

# Geo
geopy==1.22.0